Update
Update vue-intermediate as Module
