+++ vue-todo/src/components/TodoInput.vue
@@ -40,7 +40,7 @@ export default {
         }
     },
     components: {
-        Modal: Modal
+        Modal
     }
 }
 </script>
